Change Playerdata Version to Int

+++ player.h
@@ -15,7 +15,7 @@ namespace adr {
     class Player {
     private:
         // the savedata version, so that the data can be converted if needed
-        short m_version{};
+        int m_version{};
         const dpp::snowflake m_uuid{};
 
         adr::Job::Id m_job{ adr::Job::MAX };
